Rename mod metrics -> storage
